explicitly scope ::close and ::dup
Summary:
This is effectively a NOP at this point in the diff stack.
Later, when we introduce the hard folly dep, folly takes the somewhat
aggressive step of importing its implementations of these functions
into scope with a global `using` statement, and that causes this
module to fail to compile due to ambiguous symbol resolution.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9396881

fbshipit-source-id: e0a5cbe54cadcc0c0968193272d5c13d49c135ba