--watchman-path now sets WATCHMAN_BINARY
Summary:
This makes sure that tests can find the watchman binary.  Previously
all flavors of the build would contrive to place the binary at the root
of the watchman project.   Later, the windows build placed it in a `Release`
or `Debug` subdir.   While iterating on the cmake build, the binary can end
up somewhere that isn't otherwise in the path.

This diff makes sure to connect the use of the path option to setting up
the environment so that the rest of the test infra can find it.

Refs: https://github.com/facebook/watchman/pull/639

Reviewed By: strager

Differential Revision: D9552412

fbshipit-source-id: c5133f8421d98ce4f549e600fba8660484703924