adopt new thrift streaming API
Summary:
In heap profiles we observed a lot of objects associated
with the journal related to the subscription path.   Those objects
appear to be alive for the duration of the subscription session,
so it gives us good reason to move forwards with updating to the
newer thrift streaming API.

This diff introduces a new endpoint `subscribeStream` that uses
the new RSocket based subscription channel.

Both the Eden server and the watchman client are able to deal
with connections from/to old versions of the server which checks
off the box around push safety.

Once this is widely deployed we can remove `StreamingSubscriber.{cpp|h}`
from the eden code base, and can remove the `legacySubscribe` code
from the watchman code base.

Reviewed By: strager

Differential Revision: D9595967

fbshipit-source-id: 0843e56315f83f1e5fb9bc827b7ee6cf1bf507a6