use buck python unittests for integration tests
Summary:
when we first migrated the source of truth for watchman from github to
fbsource, testpilot/buck weren't able to do as good a job at managing
concurrency of the watchman tests as our own test runner.

That resulted in a rather clunky runtests wrapper script being built and run as
a custom unit test.

One of the main downsides of this approach is that it makes it more difficult
to use the normal buck/testpilot test filtering which in turn made it rather
painful to iterate and track down the recent issue with the -oss flavor of Eden
+ rsocket.

This prompted me to take a pass at making buck run the tests in the usual way;
it turns out that the runner has improved and is reasonably able to run the
watchman tests now.

Reviewed By: chadaustin

Differential Revision: D10137837

fbshipit-source-id: de2cfbf786a9f189ebb968c505c781ce2d7fffaf