Make pywatchman_aio handle connection death
Summary:
Make it so the client detects if the Watchman server connection is broken and propagates this to consumers. Propagation happens by stuffing an exception into the queues consumers are reading from, and then turning this back into an exception at the other end.

The only change for existing users is Watchman connection failure will now cause an exception. This seems strictly better than the previous behavior of quietly just not working anymore.

Reviewed By: strager

Differential Revision: D10209724

fbshipit-source-id: e84805451401c6aa2a75f589655d16c7da930f40