Back out "keep named pipe instance alive for longer"
Summary:
We've traced a hang-on-start regression in sandcastle
to this change.  The original purpose of this change was to minimize
flakeyness on CI so we can tolerate undoing this for now while we
figure out how best to solve the core problem.

This is a partial back out of 43f886849855; I'm preserving the helper
function but removing the extra unused server pipe.

https://bugzilla.mozilla.org/show_bug.cgi?id=1501416

Reviewed By: strager

Differential Revision: D12971374

fbshipit-source-id: db585ee4ccc706718294ee14ed912fb8191eb091