change Eden detection to look for FUSE filesystem type
Summary:
Update the `detectEden()` function to check that this is a FUSE filesystem,
and skip trying to make a thrift call to Eden.

Older versions of Eden had a bug where resolving the `.eden/socket` file could
temporarily return `ESTALE` errors in some situation.  This has since been
fixed (D12954819), however it is still nice to have watchman work correctly if
users are still running an older Eden version.  Previously watchman would
incorrectly fall back to trying to use an inotify watcher, which then causes
it to respond successfully to queries but not actually detect filesystem
changes.

With this new behavior Watchman should detect the watcher type correctly.  It
may still fail some queries with errors if it gets ESTALE errors trying to
connect to the Eden socket.  However, this is much better than responding
successfully with inaccurate results.

Reviewed By: wez

Differential Revision: D13336815

fbshipit-source-id: 30b855de9fb6046e4a8beccfd1a8c13ff9771c59